+++ Models/ApplicationSettingsModel.cs
@@ -156,6 +156,9 @@ namespace ThreadPilot.Models
         [ObservableProperty]
         private bool applyPersistentRulesOnProcessStart = true;
 
+        [ObservableProperty]
+        private bool enableAutomationMonitoring = true;
+
         // Advanced Settings
         [ObservableProperty]
         private bool enableDebugLogging = false;
@@ -243,6 +246,7 @@ namespace ThreadPilot.Models
             this.EnableWmiMonitoring = other.EnableWmiMonitoring;
             this.EnableFallbackPolling = other.EnableFallbackPolling;
             this.ApplyPersistentRulesOnProcessStart = other.ApplyPersistentRulesOnProcessStart;
+            this.EnableAutomationMonitoring = other.EnableAutomationMonitoring;
 
             // Advanced Settings
             this.EnableDebugLogging = other.EnableDebugLogging;